@@ -3,17 +3,17 @@
   "description": "EthicalJobs API client and SDK",
   "license": "MIT",
   "require": {
-    "php": "^7.3",
-    "guzzlehttp/guzzle": "~6.0 || ~7.0",
-    "ethical-jobs/laravel-storage": "^v2.0.0",
-    "ext-json": "*"
+    "php": "~8.0",
+    "ext-json": "*",
+    "guzzlehttp/guzzle": "^7.4"
   },
   "require-dev": {
-    "phpunit/phpunit": "^9.0",
-    "mockery/mockery": "^1.4",
-    "codedungeon/phpunit-result-printer": "^0.30",
-    "illuminate/support": "^8.34",
-    "orchestra/testbench": "^6.0"
+    "phpunit/phpunit": "^9.5",
+    "mockery/mockery": "^1.5",
+    "orchestra/testbench": "^7.6",
+    "orchestra/database": "*",
+    "roave/security-advisories": "dev-latest",
+    "laravel/legacy-factories": "^1.3"
   },
   "autoload": {
     "psr-4": {
