@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace EthicalJobs\SDK;
 
 use EthicalJobs\SDK\Authentication\Authenticator;
@@ -16,59 +18,20 @@ use GuzzleHttp\Psr7\Response;
  */
 class HttpClient
 {
-    /**
-     * Guzzle client
-     *
-     * @var Client
-     */
-    protected $client;
+    protected bool $authenticate = false;
+    protected Request $request;
+    protected Response $response;
 
-    /**
-     * Guzzle client
-     *
-     * @var Authenticator
-     */
-    protected $authenticator;
-
-    /**
-     * Authenticate requests
-     *
-     * @var bool
-     */
-    protected $authenticate = false;
-
-    /**
-     * PSR7 request
-     *
-     * @var Request
-     */
-    protected $request;
-
-    /**
-     * PSR7 response
-     *
-     * @var Response
-     */
-    protected $response;
-
-    /**
-     * Object constructor
-     *
-     * @param Client $client
-     * @param Authenticator $auth
-     */
-    public function __construct(Client $client, Authenticator $auth = null)
+    public function __construct(protected Client $client, protected ?Authenticator $authenticator = null)
     {
-        $this->client = $client;
-        $this->authenticator = $auth;
     }
 
     /**
      * Set authentication to true
      *
-     * @return HttpClient
+     * @return $this
      */
-    public function authenticate(): HttpClient
+    public function authenticate(): self
     {
         $this->authenticate = true;
 
@@ -80,11 +43,11 @@ class HttpClient
      *
      * @param string $route
      * @param array $body
-     * @param array $headers
+     * @param array<string, string> $headers
      * @return Collection
      * @throws GuzzleException
      */
-    public function get(string $route, $body = [], $headers = [])
+    public function get(string $route, array $body = [], array $headers = [])
     {
         return $this->request('GET', $route, $body, $headers);
     }
@@ -95,11 +58,11 @@ class HttpClient
      * @param string $verb
      * @param string $route
      * @param array $body
-     * @param array $headers
+     * @param array<string, string> $headers
      * @return Collection
      * @throws GuzzleException
      */
-    public function request(string $verb, string $route, $body = [], $headers = []): Collection
+    public function request(string $verb, string $route, array $body = [], array $headers = []): Collection
     {
         $request = $this->createRequest($verb, $route, $body, $headers);
 
@@ -116,10 +79,10 @@ class HttpClient
      * @param string $verb
      * @param string $route
      * @param array $body
-     * @param array $headers
+     * @param array<string, string> $headers
      * @return Request
      */
-    protected function createRequest(string $verb, string $route, $body = [], $headers = [])
+    protected function createRequest(string $verb, string $route, array $body = [], array $headers = [])
     {
         $url = Router::getRouteUrl($route);
 
@@ -135,12 +98,12 @@ class HttpClient
     }
 
     /**
-     * Merges degfault headers with user defined headers
+     * Merges default headers with user defined headers
      *
-     * @param array $headers
-     * @return array
+     * @param array<string, string> $headers
+     * @return array<string, string>
      */
-    protected function mergeDefaultHeaders(array $headers = [])
+    protected function mergeDefaultHeaders(array $headers = []): array
     {
         return array_merge([
             'Content-Type' => 'application/json',
@@ -218,15 +181,16 @@ class HttpClient
     }
 
     /**
-     * Prases a response and returns a collection or item
+     * Parses a response and returns a collection or item
      *
      * @param Response $response
      * @return Collection
      */
     protected function parseResponse(Response $response)
     {
-        if ($response->getStatusCode() > 199 && $response->getStatusCode() < 299) {
-            if ($body = json_decode($response->getBody()->getContents(), 1)) {
+        $status = $response->getStatusCode();
+        if ($status > 199 && $status < 299) {
+            if ($body = json_decode($response->getBody()->getContents(), true)) {
                 return new Collection($body);
             }
         }
@@ -239,11 +203,11 @@ class HttpClient
      *
      * @param string $route
      * @param array $body
-     * @param array $headers
+     * @param array<string, string> $headers
      * @return Collection
      * @throws GuzzleException
      */
-    public function post(string $route, $body = [], $headers = [])
+    public function post(string $route, array $body = [], array $headers = [])
     {
         return $this->request('POST', $route, $body, $headers);
     }
@@ -253,11 +217,11 @@ class HttpClient
      *
      * @param string $route
      * @param array $body
-     * @param array $headers
+     * @param array<string, string> $headers
      * @return Collection
      * @throws GuzzleException
      */
-    public function put(string $route, $body = [], $headers = [])
+    public function put(string $route, array $body = [], array $headers = [])
     {
         return $this->request('PUT', $route, $body, $headers);
     }
@@ -267,11 +231,11 @@ class HttpClient
      *
      * @param string $route
      * @param array $body
-     * @param array $headers
+     * @param array<string, string> $headers
      * @return Collection
      * @throws GuzzleException
      */
-    public function patch(string $route, $body = [], $headers = [])
+    public function patch(string $route, array $body = [], array $headers = [])
     {
         return $this->request('PATCH', $route, $body, $headers);
     }
@@ -281,11 +245,11 @@ class HttpClient
      *
      * @param string $route
      * @param array $body
-     * @param array $headers
+     * @param array<string, string> $headers
      * @return Collection
      * @throws GuzzleException
      */
-    public function delete(string $route, $body = [], $headers = [])
+    public function delete(string $route, array $body = [], array $headers = [])
     {
         return $this->request('DELETE', $route, $body, $headers);
     }
@@ -295,7 +259,7 @@ class HttpClient
      *
      * @return Request
      */
-    public function getRequest()
+    public function getRequest(): Request
     {
         return $this->request;
     }
@@ -318,7 +282,7 @@ class HttpClient
      *
      * @return Response
      */
-    public function getResponse()
+    public function getResponse(): Response
     {
         return $this->response;
     }
