@@ -1,10 +1,12 @@
 <?php
 
+declare(strict_types=1);
+
 namespace EthicalJobs\SDK\Repositories;
 
 use EthicalJobs\SDK\ApiClient;
 use EthicalJobs\SDK\Collection;
-use EthicalJobs\Storage\Contracts\Repository;
+use EthicalJobs\SDK\Storage\Contracts\Repository;
 use Illuminate\Support\Arr;
 
 /**
