@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace EthicalJobs\Tests\SDK;
 
 use EthicalJobs\SDK\Router;
@@ -12,31 +14,9 @@ class RouterTest extends TestCase
      * @test
      * @group Unit
      */
-    public function it_returns_correct_urls_for_host()
+    public function testTestEnvironmentUrlCorrect()
     {
-        putenv('API_HOST'); // omitting value removes the env var
-        putenv('API_SCHEME');
-        $this->assertEquals('https://api.ethicaljobs.com.au/jobs', Router::getRouteUrl('jobs'));
-
-        putenv('API_HOST');
-        putenv('API_SCHEME=https');
-        $this->assertEquals('https://api.ethicaljobs.com.au/jobs', Router::getRouteUrl('jobs'));
-
-        putenv('API_HOST=api.ethicaljobs.com.au');
-        putenv('API_SCHEME');
-        $this->assertEquals('https://api.ethicaljobs.com.au/jobs', Router::getRouteUrl('jobs'));
-
-        putenv('API_HOST=api.ethicaljobs.com.au');
-        putenv('API_SCHEME=https');
-        $this->assertEquals('https://api.ethicaljobs.com.au/jobs', Router::getRouteUrl('jobs'));
-
-        putenv('API_HOST=api.ethicalstaging.com.au');
-        putenv('API_SCHEME=https');
-        $this->assertEquals('https://api.ethicalstaging.com.au/jobs', Router::getRouteUrl('jobs'));
-
-        putenv('API_HOST=api-local');
-        putenv('API_SCHEME=http');
-        $this->assertEquals('http://api-local/jobs', Router::getRouteUrl('jobs'));
+        self::assertEquals('http://fail-connection-plox/jobs', Router::getRouteUrl('jobs'));
     }
 
     /**
@@ -48,11 +28,11 @@ class RouterTest extends TestCase
         App::shouldReceive('environment')->andReturn('production');
 
         $this->assertEquals(
-            'https://api.ethicalstaging.com.au/route/to/jobs',
+            'http://fail-connection-plox/route/to/jobs',
             Router::getRouteUrl('route/to/jobs')
         );
         $this->assertEquals(
-            'https://api.ethicalstaging.com.au/route/to/jobs',
+            'http://fail-connection-plox/route/to/jobs',
             Router::getRouteUrl('/route/to/jobs')
         );
     }
